Support better error logging in the HTTP service.
I found cases today where we are panicing in the app but the logs
go unnoticed because the logs are split between stackdriver lines.

This includes a panic capure to the template handler and api handler.

It includes basic error logging for cases where we return a 5xx to the
user.