APP-1068 - Add subject and kid support in JWT (#115)
Distinguish between aud (audience) and sub (subject) claims. Audience should generally be used to indicate what the authentication token is used against. While the sub indicates who the owner of the token is. Eg. user entity.

By default simple auth handlers and Authenticate request only accept an entity which is used as both the audience and the subject.

However AuthenticateTo now uses the entity of the original Authenticate request as the subject and the audience as the entity its Authenticating To.

Add SHA-1 thumbprint as the kid to all JWTs created by the RPC framework.

Background on JWT claim usage: See: https://www.rfc-editor.org/rfc/rfc7519#section-4.1