fix: Allow 10m for a user to login through google/auth0 (#126)
Currently it is set at 1 minute which does not give enough time to some users. Users may get side tracked during the login flow or take a while to sign up with google + two factor login.

This increases the time to 10m.

However we should also gracefully handle the error in a further fix like redirecting to the /login handler again.